+++ R/data.R
@@ -491,7 +491,7 @@
 #' Water quality data from the United States Bureau of Reclamation Sacramento Deepwater Ship Channel cruises.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 904 rows and 13 variables
+#' @format a tibble with 904 rows and 14 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
@@ -502,10 +502,11 @@
 #'   \item{Depth}{Bottom depth (m). Only 1 value per station, probably an average?}
 #'   \item{Sample_depth_surface}{Depth (m) of surface sample.}
 #'   \item{Sample_depth_bottom}{Depth (m) of bottom sample.}
-#'   \item{Chlorophyll}{Chlorophyll concentration (\eqn{\mu}g \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}).}
 #'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Temperature_bottom}{Temperature (°C) at bottom.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{Conductivity_bottom}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at bottom.}
+#'   \item{Chlorophyll}{Chlorophyll concentration (\eqn{\mu}g \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}).}
 #'   }
 #' @seealso \code{\link{wq}}
 "USBR"