+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Added `DOP` water quality data.
 * Updated baystudy, DJFMP, EDSM, FMWT, NCRO, SKT, SLS, STN, suisun, and twentymm datasets.
-* Added bottom conductivity to `EDSM` and `SLS`
+* Added bottom conductivity to `EDSM`, `SLS`, and `USBR`
 * Added turbidity to `DJFMP`, `EDSM`, `SKT`, `SLS`, and `twentymm`
 * Improved backend of data processing to add automation and data checks